Add `SNPsCollection` class and refactor using inheritance
